Implement eslint with husky precommit hook & lint-staged
