+++ controllers/admin.js
@@ -1,6 +1,6 @@
 let callbackObject = {
-	a: '', // Action
-	p: [], // Parameters
+  a: '', // Action
+  p: [], // Parameters
 };
 
 /**
@@ -9,39 +9,39 @@ let callbackObject = {
  * @returns {Object} Inline Keyboard
  */
 let verify = (msg) => {
-	callbackObject.action = 'v';
-	callbackObject.p[0] = msg.from.id;
-	callbackObject.p[1] = msg.message_id;
-	const inlineKeyboard = {
-		inline_keyboard: [
-			[
-				{ text: 'Haga clic aquí', callback_data: JSON.stringify(callbackObject) },
-			],
-		],
-	};
-	return (inlineKeyboard);
+  callbackObject.action = 'v';
+  callbackObject.p[0] = msg.from.id;
+  callbackObject.p[1] = msg.message_id;
+  const inlineKeyboard = {
+    inline_keyboard: [
+      [
+        { text: 'Haga clic aquí', callback_data: JSON.stringify(callbackObject) },
+      ],
+    ],
+  };
+  return (inlineKeyboard);
 };
 
 const RemPerms = {
-	perms: {
-		can_send_message: false,
-		can_send_media_messages: false,
-		can_send_other_messages: false,
-		can_add_web_page_previews: false,
-	},
+  perms: {
+    can_send_message: false,
+    can_send_media_messages: false,
+    can_send_other_messages: false,
+    can_add_web_page_previews: false,
+  },
 };
 
 const GivePerms = {
-	perms: {
-		can_send_message: true,
-		can_send_media_messages: false,
-		can_send_other_messages: false,
-		can_add_web_page_previews: false,
-	},
+  perms: {
+    can_send_message: true,
+    can_send_media_messages: false,
+    can_send_other_messages: false,
+    can_add_web_page_previews: false,
+  },
 };
 
 module.exports = {
-	verify,
-	RemPerms,
-	GivePerms,
+  verify,
+  RemPerms,
+  GivePerms,
 };
