+++ models/user.js
@@ -2,8 +2,8 @@ let mongoose = require('mongoose');
 let Schema = mongoose.Schema;
 
 let userSchema = new Schema({
-	userId: { type: Number, index: { unique: true } },
-	username: {type: String},
+  userId: { type: Number, index: { unique: true } },
+  username: {type: String},
 }, { versionKey: false });
 
 let userModel = mongoose.model('User', userSchema);
